Improved the line by line render behaviour
The line by line render behaviour is now correct, and displays
correctly. The previous black line artifacts have been corrected.

The program now records the colour of each point in the plot.
These are then rendered when the render() function is called.

If render is called on a line by line basis then there will be a
significant performance impact.