Updated Mandelbrot that uses the toolbox
Updated the mandelbrot program so that it uses the toolbox
for the graphics. This simplifies the program.
The program now includes timing information.

The working program is called mandelbrot-complete.go. The students
should complete the mandelbrot.go program.

A updated and working example of the Julia set is in
julia-complete.go